category form

+++ db/migrate/20150903195733_create_event_expense_categories.rb
@@ -3,7 +3,7 @@ class CreateEventExpenseCategories < ActiveRecord::Migration
     create_table :event_expense_categories, id: :uuid do |t|
       t.uuid :event_id
       t.uuid :expense_category_id
-      t.decimal :budget
+      t.float :budget
 
       t.timestamps null: false
     end
